Rename sh3_arc_file to sh3::arc::mft & use Doxygen comments
