+++ source/CMakeLists.txt
@@ -19,7 +19,7 @@ include_directories(SYSTEM "${ZLIB_INCLUDE_DIRS}")
 add_executable("${CMAKE_PROJECT_NAME}"
 	"main.cpp"
 	
-	"SH3/arc/file.cpp"
+	"SH3/arc/mft.cpp"
 	"SH3/arc/section.cpp"
 	"SH3/arc/types.cpp"
 	"SH3/arc/vfile.cpp"
