+++ tests/CMakeLists.txt
@@ -56,7 +56,7 @@ target_link_libraries("vao"
 add_executable("tex"
 	"tex.cpp"
 	
-	"../source/SH3/arc/file.cpp"
+	"../source/SH3/arc/mft.cpp"
 	"../source/SH3/arc/section.cpp"
 	"../source/SH3/arc/types.cpp"
 	"../source/SH3/arc/vfile.cpp"
